+++ WebSocketSource.cs
@@ -52,7 +52,7 @@ namespace ArkeIndustries.RequestServer.Sources {
 				return true;
 			}
 
-			protected override async Task<int> Receive(MemoryStream stream, long offset, long length) {
+			protected override async Task<long> Receive(MemoryStream stream, long offset, long length) {
 				var result = await this.client.ReceiveAsync(new ArraySegment<byte>(stream.GetBuffer(), (int)offset, (int)length), CancellationToken.None);
 
 				return result.MessageType == WebSocketMessageType.Binary ? result.Count : 0;
